Fix incorrect pixel values in empty latent error message
Example: a valid image size of 960x544 would show an error message of: "Latent dimensions must be divisible by 16. Got 68x120. For a 544x960 pixel image, use a 64x112 latent". The bug is because the original code used height and width values from the shape object to populate error messages, but these are NOT pixel values because they have undergone  //8 for latent compression. So the error message values were incorrect and confusing (and had height, width rather than the conventional order of width, height).